Add test for PlaidIndexer and PlaidSearcher functionality
